rename GlobalEntityControl to GlobalEntitySelector
