--- grapher/controls/globalEntityControl/GlobalEntityRegistry.ts
@@ -1,3 +1,0 @@
-import { SelectionArray } from "../../selection/SelectionArray"
-
-export const GlobalEntityRegistry = new Set<{ selection: SelectionArray }>()
