+++ grapher/controls/globalEntitySelector/GlobalEntitySelector.scss
