+++ grapher/controls/globalEntitySelector/GlobalEntitySelector.stories.tsx
@@ -1,12 +1,12 @@
 import * as React from "react"
 import { SelectionArray } from "../../selection/SelectionArray"
-import { GlobalEntityControl } from "./GlobalEntityControl"
+import { GlobalEntitySelector } from "./GlobalEntitySelector"
 
 export default {
     title: "GlobalEntityControl",
-    component: GlobalEntityControl,
+    component: GlobalEntitySelector,
 }
 
 export const WithNoGraphers = () => (
-    <GlobalEntityControl selection={new SelectionArray()} />
+    <GlobalEntitySelector selection={new SelectionArray()} />
 )
