+++ grapher/controls/globalEntitySelector/GlobalEntitySelectorConstants.ts
