+++ grapher/controls/globalEntitySelector/readme.md
