+++ grapher/core/grapher.scss
@@ -132,7 +132,7 @@ $zindex-Tooltip: 20;
 }
 
 @import "grapher/controls/entityPicker/EntityPicker.scss";
-@import "grapher/controls/globalEntityControl/GlobalEntityControl.scss";
+@import "grapher/controls/globalEntitySelector/GlobalEntitySelector.scss";
 @import "grapher/controls/CommandPalette.scss";
 @import "grapher/controls/ScaleSelector.scss";
 @import "grapher/controls/Controls.scss";
