+++ site/multiembedder/MultiEmbedder.tsx
@@ -19,11 +19,11 @@ import {
     GLOBAL_ENTITY_CONTROL_DATA_ATTR,
     GLOBAL_ENTITY_CONTROL_DEFAULT_COUNTRY,
     GLOBAL_ENTITY_CONTROL_SELECTOR,
-} from "../../grapher/controls/globalEntityControl/GlobalEntityControlConstants"
+} from "../../grapher/controls/globalEntitySelector/GlobalEntitySelectorConstants"
 import { Url } from "../../urls/Url"
 import { SelectionArray } from "../../grapher/selection/SelectionArray"
 import { EntityUrlBuilder } from "../../grapher/core/EntityUrlBuilder"
-import { hydrateGlobalEntityControlIfAny } from "../../grapher/controls/globalEntityControl/GlobalEntityControl"
+import { hydrateGlobalEntityControlIfAny } from "../../grapher/controls/globalEntitySelector/GlobalEntitySelector"
 
 interface EmbeddedFigureProps {
     standaloneUrl: string
